Version 0.2.0.0
This is the first version to use version >= 0.6 of language-javascript.

This new version of language-javascript has a completely different
AST (abstract syntax tree) from versions < 0.6 and includes a module
which does minification.

This version of hjsmin is therefore a thin wrapper around language-javascript
that does no more than parse command lines argumentx, determinine input and
output file names and do the required plumbing.